requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask~=3.0.0
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2~=3.1.2
MarkupSafe==2.1.3
typing_extensions~=4.8.0
Werkzeug==3.0.1

python-dotenv~=1.0.0
pymongo~=4.6.1